Move terminal binding and policy behavior to helper
